requirements.txt
streamlit
python-dotenv
numpy
pandas
scikit-learn



google-generativeai

